requirements.txt
flask
flask_sqlalchemy
flask_migrate
pyodbc
python-dotenv
